feat(FiltersBar) moved to separate component and used icons from figma
